Pretend Qt5 wasn't found if OE_QMAKE_PATH_EXTERNAL_HOST_BINS isn't set
This prevents errors like:

|   The imported target "Qt5::Core" references the file
|
|      "/qmake"
|
|   but this file does not exist.  Possible reasons include:

Which happen if CMake is used without setting OE_QMAKE_PATH_EXTERNAL_HOST_BINS.
To achieve this a check for OE_QMAKE_PATH_EXTERNAL_HOST_BINS is added to each Qt5*Config.cmake
file. And in the case where the variable is not set we just return which is basically
equal to telling CMake that Qt5 wasn't found.

Upstream-Status: Pending
  The patch only makes sense in connection with other patches included here.
  Specifically this are:
    - 0003-Add-external-hostbindir-option.patch
    - 0010-Add-external-hostbindir-option-for-native-sdk.patch

Signed-off-by: Pascal Bach <pascal.bach@siemens.com>